+++ CHANGELOG.md
@@ -1,6 +1,15 @@
 # Changelogs for IADM
 
 
+## v0.4 beta4 (#17)
++ Status command now also displays on which operating system the server runs on.
++ Added monitorlogs command, toggles monitoring for displaying any new logs into the console.
+
+/ Changed viewfulllogs command functionality a little bit
+
+
+
+
 ## Fixes (#16)
 + Added IADM:ProcessArgs function